# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: nicksan222@icloud.com
    name: API Support
  description: This is the KetoAI API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: KetoAI API
  version: "1.0"
paths:
  /ingredients:
    get:
      description: Retrieves a list of ingredients based on beginning and ending
        characters and a limit on the number of results.
      parameters:
        - description: Filter ingredients that begin with these characters
          in: query
          name: begins_with
          schema:
            type: string
        - description: Filter ingredients that end with these characters
          in: query
          name: ends_with
          schema:
            type: string
        - description: Limit the number of ingredients returned
          in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of ingredients returned successfully.
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ingredients_list.ListIngredientsResponse"
                type: array
        "400":
          description: Bad Request - Invalid query parameters.
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Not Found - Ingredients not found.
          content:
            application/json:
              schema:
                type: object
      summary: List ingredients
      tags:
        - ingredients
  "/ingredients/{ingredient_id}":
    get:
      description: Retrieves the ingredient by its ID.
      parameters:
        - description: Ingredient ID
          in: path
          name: ingredient_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ingredient found and returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ingredients_get.GetIngredientResponse"
        "400":
          description: Bad Request - Missing ingredient ID.
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Not Found - Ingredient not found.
          content:
            application/json:
              schema:
                type: object
      summary: Retrieve a single ingredient
      tags:
        - ingredients
  /ingredients/preferences:
    get:
      description: Retrieves the list of ingredient preferences for a user based on
        their ID.
      responses:
        "200":
          description: List of ingredient preferences returned successfully.
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ingredients_getpreferences.GetIngredientPreferences\
                    Response"
                type: array
        "401":
          description: Unauthorized - User ID not found.
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Not Found - Ingredient preferences not found.
          content:
            application/json:
              schema:
                type: object
      summary: Retrieve a user's ingredient preferences
      tags:
        - ingredients
    post:
      description: Sets or updates the ingredient preferences for a user based on their ID.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ingredients_setpreferences.SetIngredientPreferences\
                Request"
        description: Ingredient Preferences Request
        required: true
      responses:
        "200":
          description: Ingredient preferences set or updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ingredients_setpreferences.SetIngredientPreferences\
                  Response"
        "400":
          description: Bad Request - Invalid request body or parameters.
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized - User ID not found.
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Not Found - Unable to set or update preferences.
          content:
            application/json:
              schema:
                type: object
      summary: Set a user's ingredient preferences
      tags:
        - ingredients
  "/ingredients/preferences/{ingredient_id}":
    delete:
      description: Deletes the preference of a specific ingredient for a user based on
        their ID and the ingredient ID.
      parameters:
        - description: Ingredient ID
          in: path
          name: ingredient_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ingredient preference deleted successfully.
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Bad Request - Missing ingredient ID.
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized - User ID not found.
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Not Found - Ingredient or preference not found.
          content:
            application/json:
              schema:
                type: object
      summary: Delete a user's ingredient preference
      tags:
        - ingredients
servers:
  - url: //http://backend:4000
components:
  schemas:
    ingredients.Ingredient:
      properties:
        approved:
          description: Users can insert custom ingredients, but must be approved first by
            admins
          type: boolean
        carbs:
          type: number
        fat:
          type: number
        id:
          type: string
        is_main_ingredient:
          description: Main ingredients are used in the calculation of the macros of a
            recipe
          type: boolean
        name:
          type: string
        protein:
          type: number
        quantity_measurement:
          type: string
      type: object
    ingredients_get.GetIngredientResponse:
      properties:
        ingredient:
          $ref: "#/components/schemas/ingredients.Ingredient"
      type: object
    ingredients_getpreferences.GetIngredientPreferencesResponse:
      properties:
        ingredients:
          items:
            type: string
          type: array
        user_id:
          type: string
      type: object
    ingredients_list.ListIngredientsResponse:
      properties:
        ingredients:
          items:
            $ref: "#/components/schemas/ingredients.Ingredient"
          type: array
      type: object
    ingredients_setpreferences.SetIngredientPreferencesRequest:
      properties:
        ingredient_ids:
          items:
            type: string
          type: array
        user_id:
          type: string
      type: object
    ingredients_setpreferences.SetIngredientPreferencesResponse:
      properties:
        ingredient_ids:
          items:
            type: string
          type: array
      type: object